requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
pillow>=9.5.0
scikit-learn>=1.3.0
tqdm>=4.65.0
networkx>=3.1

# Google Sheets export (optional — only needed if using sheets_config.json)
gspread>=6.0.0
google-auth>=2.0.0